requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow
numpy
Pillow
vercel-blob
